+++ externs.js
@@ -12,26 +12,26 @@
  */
 
 /**
- * This class provides a div with terminal functionalities
+ *  This class provides a div with terminal functionalities
  *
- * @id 20190208°1923`12
+ * @id 20190208°1923'02
  * @type {Object}
  */
 var Terminal;
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0521`12
+ * @id 20170501°0521'02
  * @type {Function} —
  * @_return {undefined} —
  */
 Terminal.beep();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0531`12
+ * @id 20170501°0531'02
  * @type {Function} —
  * @param {boolean} —
  * @_return {undefined} —
@@ -39,18 +39,18 @@ Terminal.beep();
 Terminal.blinkingCursor();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0541`12
+ * @id 20170501°0541'02
  * @type {Function} —
  * @_return {undefined} —
  */
 Terminal.clear();
 
 /**
- * Clear the history
+ *  Clear the history
  *
- * @id  20210503°0931`12
+ * @id  20210503°0931'02
  * @type {Function} —
  * @_return {undefined} —
  */
@@ -62,7 +62,7 @@ Terminal.clearHistory();
       | /**
       |  * @type {Function} —
       |  * @param {Function} callback —
-      | * /
+      |  * /
       the GoCloCom emits 'WARNING - [JSC_TYPE_PARSE_ERROR] Bad type annotation'
    see : https://github.com/google/closure-compiler/issues/3767 [ref 20210430°1532] not so helpful
    workaround : Keep the offending parts outcommented
@@ -72,9 +72,9 @@ Terminal.clearHistory();
 */
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0551`12
+ * @id 20170501°0551'02
  * @type {Function} —
  * @note Remember issue 20210430°1531 'JSC_TYPE_PARSE_ERROR'
  * @param {string} // [message] —
@@ -84,18 +84,18 @@ Terminal.clearHistory();
 Terminal.confirm();
 
 /**
- * Switch on XHR mode and provide backend address
+ *  Switch on XHR mode and provide backend address
  *
- * @id 20210502°1211`12
+ * @id 20210502°1211'02
  * @param {string} url —
  * @return {undefined} —
  */
 Terminal.connect();
 
 /**
- * Get ID of instance
+ *  Get ID of instance
  *
- * @id 20210509°1631`12
+ * @id 20210509°1631'02
  * @_t_y_p_e {Function} —
  * @return {string} —
  */
@@ -104,23 +104,23 @@ Terminal.getId();
 /**
  *  Returns the version string of the program
  *
- * @id 20210513°0913'12
+ * @id 20210513°0913'02
  * @return {string} —
  */
 Terminal.getVersion;
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0411`12
+ * @id 20170501°0411'02
  * @type {Node} —
  */
 Terminal.html;
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0611`12
+ * @id 20170501°0611'02
  * @type {Function} —
  * @param {string} // message —
  * @param {Function} // callback —
@@ -129,9 +129,9 @@ Terminal.html;
 Terminal.input();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0621`12
+ * @id 20170501°0621'02
  * @type {Function} —
  * @param {string} // message —
  * @param {Function} // callback —
@@ -140,18 +140,18 @@ Terminal.input();
 Terminal.password();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0631`12
+ * @id 20170501°0631'02
  * @param {string} —
  * @type {Function} —
  */
 Terminal.print();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0641`12
+ * @id 20170501°0641'02
  * @param {string} —
  * @type {Function} —
  */
@@ -169,9 +169,9 @@ Terminal.setBackgroundColor();
 //this.setDebugBorders();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0651`12
+ * @id 20170501°0651'02
  * @param {string} —
  * @type {Function} —
  */
@@ -180,7 +180,7 @@ Terminal.setHeight();
 /**
  *  Lets the user set an input prompt, e.g. '$ '.
  *
- * @id 20210504°1011`12
+ * @id 20210504°1011'02
  * @param {string} — The wanted input prompt
  * @type {Function} —
  */
@@ -189,43 +189,52 @@ Terminal.setInputPrompt();
 /**
  *  Lets the user set an output prompt, e.g. '=> ' -- Not yet used
  *
- * @id 20210504°1021`12
+ * @id 20210504°1021'02
  * @param {string} — The wanted output prompt
  * @type {Function} —
  */
 Terminal.setOutputPrompt();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0711`12
+ * @id 20170501°0711'02
  * @param {string} —
  * @type {Function} —
  */
 Terminal.setTextColor();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0721`12
+ * @id 20170501°0721'02
  * @param {string} —
  * @type {Function} —
  */
 Terminal.setTextSize();
 
 /**
- * ..
+ *  Set beep volume. Takes a value in the range from 0.0 to 1.0
  *
- * @id 20170501°0731`12
+ * @id 20210513°1621'02
+ * @param {string} —
+ * @type {Function} —
+ */
+Terminal.setVolume();
+
+/**
+ *  ..
+ *
+ * @id 20170501°0731'02
  * @param {string} —
  * @type {Function} —
  */
 Terminal.setWidth();
 
 /**
- * ..
+ *  ..
  *
- * @id 20170501°0741`12
+ * @id 20170501°0741'02
  * @type {Function} —
  */
 Terminal.sleep();